Move test files to match new structure
